Move bindings resolving into the Actor
